refactor : Initializer 클래스 DomainsInitializer로 rename 및 리팩토링
