+++ src/main/java/lotto/model/LottoRank.java
